disable circular deps warning until forwardRef is recognized; move directive into its own component to satisfy nglint;
