+++ lib/src/radio-button/radio-button.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RadioButtonComponent, RadioGroupDirective } from './radio-button.component';
+import { RadioButtonComponent } from './radio-button.component';
+import { RadioGroupDirective } from './radio-group.directive';
 @NgModule({
     imports: [
         CommonModule
